# wa-openapi-grc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "explanation-date",
    "description": "API generated by SwaggerHub Explore",
    "version": "1.1"
  },
  "servers": [
    {
      "url": ""
    }
  ],
  "paths": {
    "/issue/details": {
      "get": {
        "parameters": [
          {
            "name": "issue_id",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            },
            "examples": {
              "example": {
                "value": "AP8876"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated by SwaggerHub Explore",
            "headers": {
              "date": {
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "example": "Wed, 20 Nov 2024 14:46:54 GMT"
              },
              "server": {
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "example": "uvicorn"
              },
              "content-length": {
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "example": "240"
              },
              "content-type": {
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "example": "application/json"
              },
              "strict-transport-security": {
                "style": "simple",
                "explode": false,
                "schema": {
                  "type": "string"
                },
                "example": "max-age=63072000; preload"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                },
                "example": {
                  "dev_ti": "Sem plano de ação",
                  "discipline": "SOx",
                  "novo_nivelrisco": "Elevado",
                  "discussion_date": "08/11/2024",
                  "data_maxima_para_implementacao": "14/11/2024",
                  "data_limite_excecao_gov": "05/11/2024"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
      "securitySchemes": {
        "BasicAuthCredentials": {
          "type": "http",
          "scheme": "basic"
        }
      },
    "schemas": {
      "inline_response_200": {
        "required": [
          "data_limite_excecao_gov",
          "data_maxima_para_implementacao",
          "dev_ti",
          "discipline",
          "discussion_date",
          "novo_nivelrisco"
        ],
        "type": "object",
        "properties": {
          "tem_cvi": {
            "type": "string"
          },
          "discipline": {
            "type": "string"
          },
          "novo_nivelrisco": {
            "type": "string"
          },
          "discussion_date": {
            "type": "string"
          },
          "data_maxima_para_implementacao": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "data_limite_excecao_gov": {
            "type": "string"
          },
          "action_item_count": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}